+++ speedd-ml/src/main/scala/org/speedd/ml/learners/CNRSWeightsEstimator.scala
@@ -31,91 +31,29 @@ package org.speedd.ml.learners
 import java.io._
 import auxlib.log.Logging
 import lomrf.logic._
-import lomrf.mln.model.{EvidenceBuilder, KB}
+import lomrf.mln.model.{ConstantsDomain, EvidenceBuilder, KB}
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{DataFrame, SQLContext}
-import org.apache.spark.sql.functions._
 import com.datastax.spark.connector._
 import org.speedd.ml.model.CNRS
 import org.speedd.ml.model.CNRS.{Location, Annotation, RawInput}
+import org.speedd.ml.util.data._
 import scala.collection.breakOut
-import java.util.Arrays.binarySearch
 import scala.language.implicitConversions
+import scala.reflect.macros.whitebox
 import scala.util.{Failure, Success}
-import org.speedd.ml.util.DataFrameOps._
 import org.speedd.ml.util.logic._
 
 object CNRSWeightsEstimator extends WeightEstimator with Logging {
-
-  private type RAW_INPUT_TUPLE = (Long, String, Int, Option[Double], Option[Double], Option[Double])
-  private type ANNOTATION_TUPLE = (Int, Int, String)
 
   private val queryPredicates = Set[AtomSignature](AtomSignature("HoldsAt", 2))
   private val hiddenPredicates = Set[AtomSignature](AtomSignature("InitiatedAt", 2), AtomSignature("TerminatedAt", 2))
   private val convertFunctionsToPredicates = false
 
-  private val inputColumnNames = Seq("loc_id", "lane", "timestamp", "avg_speed", "occupancy", "vehicles")
-
   private val tempEvidenceTableName = "TEMP_EVIDENCE"
-  private val tempLocationTableName = "TEMP_LOCATION"
 
   private val domainAliases = Map("start_loc" -> "loc_id", "end_loc" -> "loc_id")
-
-  private val occLevels = Array(0.0, 25.0, 50.0, 75.0)
-  private val speedLevels = Array(0.0, 50.0, 90.0, 100.0, 130.0)
-  private val vehicleLevels = Array(0.0, 4.0, 8.0, 16.0, 32.0)
-
-
-  private val domain2udf = Map(
-    "occupancy" -> udf {
-      (x: Double) =>
-        val pos = binarySearch(occLevels, x)
-        val bin = if (pos < 0) -pos - 2 else pos
-        "O" + bin
-    },
-
-    "avg_speed" -> udf {
-      (x: Double) =>
-        val pos = binarySearch(speedLevels, x)
-        val bin = if (pos < 0) -pos - 2 else pos
-        "S" + bin
-    },
-
-    "vehicles" -> udf {
-      (x: Double) =>
-        val pos = binarySearch(vehicleLevels, x)
-        val bin = if (pos < 0) -pos - 2 else pos
-        "V" + bin
-    }
-  )
-
-
-  implicit def string2ColumnNames(names: Seq[String]): Seq[ColumnRef] = {
-    names.map(name => new ColumnName(name))
-  }
-
-
-  private def collectDomains(source: DataFrame*): Map[String, RDD[String]] = {
-
-    def domainsOf(df: DataFrame): Map[String, DataFrame] = {
-      df.columns.map { colName =>
-        domainAliases.getOrElse(colName, colName) -> df.select(colName).filter(df(colName).isNotNull).distinct()
-      }(breakOut)
-    }
-
-    def mergeDomains(m1: Map[String, DataFrame], m2: Map[String, DataFrame]): Map[String, DataFrame] = m1 ++ m2.map {
-      case (k, v) => m1.get(k) match {
-        case Some(d) => k -> d.unionAll(v)
-        case None => k -> v
-      }
-    }
-
-    source.map(domainsOf)
-      .aggregate(Map.empty[String, DataFrame])(mergeDomains, mergeDomains)
-      .mapValues(_.distinct().map(_.get(0).toString))
-  }
-
 
   override def learn(startTime: Long, endTime: Long, inputKB: File, outputKB: File,
                      inputSignatures: Set[AtomSignature], targetSignatures: Set[AtomSignature])
@@ -143,7 +81,7 @@ object CNRSWeightsEstimator extends WeightEstimator with Logging {
 
           case Success(path) => info(s"KB transformations are written into '$path'")
 
-          case Failure(ex) => info(s"Failed to write KB transformations", ex)
+          case Failure(ex) => error(s"Failed to write KB transformations", ex)
         }
 
         // give the resulting transformations
@@ -162,120 +100,20 @@ object CNRSWeightsEstimator extends WeightEstimator with Logging {
       finalPredicateSchema ++= transformation.schema
 
 
-    // make sure that we have all time-points of the specified temporal interval
-    val timestampDomain = startTime to endTime
-    val timestampsDF = sc.parallelize(timestampDomain).toDF("timestamp").cache()
-
-    // ---
-    // --- Load raw data for the specified temporal interval
-    // ---
-    // The data is loaded from raw input
-    info(s"Loading raw data for the temporal interval [$startTime, $endTime]")
-
-    // discretize numeric values such as `occupancy`,`avg_speed`, etc.
-    val mappedInputColumns = inputColumnNames.map(colName => domain2udf.get(colName) match {
-      case Some(f) => f($"$colName") as colName
-      case None => $"$colName"
-    })
-
-    val rawInputDF = sc.cassandraTable[RAW_INPUT_TUPLE](CNRS.KEYSPACE, RawInput.tableName)
-      .select(inputColumnNames: _*) // select only the desired subset of columns, with the given ordering
-      .where("timestamp >= ? AND timestamp <= ?", startTime, endTime) // take the rows of the specified temporal interval
-      .toDF(inputColumnNames: _*) // represent the selected rows as a data frame
-      .select(mappedInputColumns: _*) // apply conversions, e.g., discretize numeric values such as `occupancy`,`avg_speed`, etc.
-      .cache() // try to cache the data frame, because we will parse it multiple times.
-
-    // Represent the selected data as a temporary in-memory table, named as `TEMP_EVIDENCE`.
-    rawInputDF.registerTempTable(tempEvidenceTableName)
-
-    // ---
-    // --- Load location data for the specified temporal interval
-    // ---
-    info(s"Loading location data")
-    val locationDF = sc.cassandraTable[Location](CNRS.KEYSPACE, Location.tableName)
-      .toDF(Location.columnNames: _*)
-      .cache()
-
-    // Represent the location data as a temporary in-memory table, named as `TEMP_LOCATION`.
-    locationDF.registerTempTable(tempLocationTableName)
-
-    // ---
-    // --- Load annotation data for the specified interval
-    // ---
-    info(s"Loading annotation data for the temporal interval [$startTime, $endTime]")
-
-    // Collect overlapping Annotations with the specified temporal interval [startTime, endTime].
-    // Thereafter, flatten the annotations from Annotation instances that represent the rows of (start_ts, end_ts,
-    // event_num, description, sens, start_loc, end_loc) to tuples of (ts, event_num, description, sens, start_loc,
-    // end_loc). Therefore we convert the Annotations containing intervals (start_ts, end_ts, ...) to multiple
-    // instantaneous annotations as tuples of (ts, ...).  The resulting columns of the `overlappingAnnotations` data
-    // frame are (ts, event_num, description, sens, start_loc, end_loc)
-    /*val overlappingAnnotationsDF = sc.cassandraTable[Annotation](CNRS.KEYSPACE, Annotation.tableName)
-      .filter(r => r.startTs <= endTime && r.endTs >= startTime)
-      .flatMap(r => (r.startTs to r.endTs).map(t => (t, r.eventNum, r.description, r.sens, r.startLoc, r.endLoc)))
-      .toDF("ts", "event_num", "description", "sens", "end_loc", "start_loc")*/
-
-
-    val l2d = locationDF
-      .select("loc_id", "dist", "lane")
-      .distinct()
-      .cache()
-
-    val annotatedLocations = sc.cassandraTable[Annotation](CNRS.KEYSPACE, Annotation.tableName)
-      .filter(r => r.startTs <= endTime && r.endTs >= startTime)
-      .toDF(Annotation.columnNames: _*)
-      .select($"start_ts", $"end_ts", $"event_num" as "eid", $"description", $"sens", $"start_loc", $"end_loc")
-      .cache()
-
-    val annotatedLocationWIds = l2d
-      .join(annotatedLocations, $"dist" <= $"start_loc" and $"dist" >= $"end_loc")
-      .select("loc_id", "start_ts", "end_ts", "description", "lane")
-      .flatMap{ r =>
-        val loc_id = r.getLong(0)
-        val description = r.getString(3)
-        val lane = r.getString(4)
-        (r.getInt(1) to r.getInt(2)).map(t => (t, loc_id, description, lane))
-      }.toDF("ts", "loc_id", "description", "lane")
-      .cache()
-
-    // We take the left outer join between `timestampsDF` and `overlappingAnnotations`, in order to collect the
-    // annotation data (including missing instances as nulls) for the specified temporal interval [startTime, endTime].
-    val annotationsDF = timestampsDF
-      .join(annotatedLocationWIds, $"timestamp" === $"ts", "left_outer")
-      .select("timestamp", "loc_id", "description", "lane")
-      .cache()
-
-    println("FINAL ANNOTATION DATA!!!!")
-    annotationsDF.show()
-
-
-   // sys.exit()
-
-
-   /* val annotationsDF = timestampsDF
-      .join(overlappingAnnotationsDF, $"timestamp" === $"ts", "left_outer")
-      .select($"timestamp", $"start_loc", $"end_loc", $"description")
-      //.join(sdistIntervals, $"start_loc" === $"startDist" and $"end_loc" === $"endDist")
-      //.join(distIntervals, $"start_loc" <= $"endDist" and $"end_loc" >= $"startDist")
-      .distinct()
-      .cache()
-
-
-    annotationsDF.show(10)
-    sys.exit()*/
+    val (timestampsDF, rawInputDF, locationDF, annotationsDF, annotatedLocations) = loadFor(startTime, endTime)
 
 
     // ---
     // --- Create constant domains:
     // ---
     //
-    val domainMap = collectDomains (
+    val domainMap = symbolsPerColumn(
       rawInputDF.select("avg_speed", "occupancy", "vehicles").distinct(),
       locationDF.select("loc_id", "lane").distinct(),
       annotatedLocations.select("description").distinct()
-    )
+    )(domainAliases)
 
-    constantsDomainBuilder ++= ("timestamp", startTime to endTime map(_.toString) )
+    constantsDomainBuilder ++=("timestamp", startTime to endTime map (_.toString))
 
     for ((domainName, constantsRDD) <- domainMap)
       constantsDomainBuilder ++=(domainName, constantsRDD.collect())
@@ -377,7 +215,7 @@ object CNRSWeightsEstimator extends WeightEstimator with Logging {
     // --- Create ground-truth predicates (HoldsAt/2)
     // ---
     info(s"Generating annotation predicates for the temporal interval [$startTime, $endTime]")
-    val headSignatures = Set(AtomSignature("InitiatedAt", 2), AtomSignature("TerminatedAt",2))
+    val headSignatures = Set(AtomSignature("InitiatedAt", 2), AtomSignature("TerminatedAt", 2))
 
     val fluents = kb.definiteClauses
       .withFilter(wc => headSignatures.contains(wc.clause.head.signature))
@@ -405,10 +243,10 @@ object CNRSWeightsEstimator extends WeightEstimator with Logging {
 
         val sterms = terms.map(_.substitute(theta))
 
-        if(sterms.forall(_.isConstant)) {
+        if (sterms.forall(_.isConstant)) {
           val fluentSignature = AtomSignature(symbol, sterms.size)
 
-          val resultingSymbolOpt = functionMappingsMap.value.get(fluentSignature).flatMap(mapper => mapper.get(sterms.map(_.toText)))
+          //val resultingSymbolOpt = functionMappingsMap.value.get(fluentSignature).flatMap(mapper => mapper.get(sterms.map(_.toText)))
 
           //resultingSymbolOpt
 
@@ -439,19 +277,19 @@ object CNRSWeightsEstimator extends WeightEstimator with Logging {
 
       val holdsAtInstances = holdsAtInstancesRDD.collect()
 
-      for(atom <- holdsAtInstances) try {
+      for (atom <- holdsAtInstances) try {
         trainingDB.evidence += atom
       } catch {
         case ex: java.util.NoSuchElementException =>
           val fluent = atom.terms.head
           val timestamp = atom.terms.last
 
-          constantsDomain("fluent").get(fluent.symbol) match{
+          constantsDomain("fluent").get(fluent.symbol) match {
             case None => error(s"fluent constant ${fluent.symbol} is missing from constants domain}")
             case _ =>
           }
 
-          constantsDomain("timestamp").get(timestamp.symbol) match{
+          constantsDomain("timestamp").get(timestamp.symbol) match {
             case None => error(s"timestamp constant ${timestamp.symbol} is missing from constants domain}")
             case _ =>
           }
@@ -465,6 +303,70 @@ object CNRSWeightsEstimator extends WeightEstimator with Logging {
 
     println("ANNOTATION length: " + edb(AtomSignature("HoldsAt", 2)).numberOfTrue)
 
+  }
+
+  private def loadFor(startTime: Long, endTime: Long)(implicit sc: SparkContext, sqlContext: SQLContext) = {
+    import sqlContext.implicits._
+
+    // make sure that we have all time-points of the specified temporal interval
+    val timestampsDF = sc.parallelize(startTime to endTime).toDF("timestamp").cache()
+
+    // ---
+    // --- Load raw data for the specified temporal interval
+    // ---
+    // The data is loaded from raw input
+    info(s"Loading raw data for the temporal interval [$startTime, $endTime]")
+
+    val rawInputDF = RawInput.loadDF
+      .where(s"timestamp >= $startTime AND timestamp <= $endTime")
+      .cache()
+
+    // Represent the selected data as a temporary in-memory table, named as `TEMP_EVIDENCE`.
+    rawInputDF.registerTempTable(tempEvidenceTableName)
+
+    // ---
+    // --- Load location data for the specified temporal interval
+    // ---
+    info(s"Loading location data")
+    val locationDF = Location.loadDF.cache()
+
+    // ---
+    // --- Load annotation data for the specified interval
+    // ---
+    info(s"Loading annotation data for the temporal interval [$startTime, $endTime]")
+
+    val annotatedLocations = Annotation.loadDF
+      .where(s"start_ts <= $endTime AND end_ts >= $startTime")
+      .withColumnRenamed("event_num", "eid")
+      .cache()
+
+    val annotationsDF = expandAnnotation(annotatedLocations, locationDF, timestampsDF).cache()
+
+    (timestampsDF, rawInputDF, locationDF, annotationsDF, annotatedLocations)
+  }
+
+  private def expandAnnotation(annotatedLocations: DataFrame, locationDF: DataFrame, timestampsDF: DataFrame)
+                              (implicit sc: SparkContext, sqlContext: SQLContext): DataFrame = {
+    import sqlContext.implicits._
+
+    val annotatedLocationWIds = locationDF.select("loc_id", "dist", "lane").distinct()
+      .join(annotatedLocations, $"dist" <= $"start_loc" and $"dist" >= $"end_loc")
+      .select("loc_id", "start_ts", "end_ts", "description", "lane")
+      .flatMap { r =>
+        val loc_id = r.getLong(0)
+        val description = r.getString(3)
+        val lane = r.getString(4)
+        (r.getInt(1) to r.getInt(2)).map(t => (t, loc_id, description, lane))
+      }
+      .toDF("ts", "loc_id", "description", "lane")
+
+    // We take the left outer join between `timestampsDF` and `overlappingAnnotations`, in order to collect the
+    // annotation data (including missing instances as nulls) for the specified temporal interval [startTime, endTime].
+    val result = timestampsDF
+      .join(annotatedLocationWIds, $"timestamp" === $"ts", "left_outer")
+      .select("timestamp", "loc_id", "description", "lane")
+
+    result
   }
 
 }
