+++ speedd-ml/src/main/scala/org/speedd/ml/learners/DataSet.scala
@@ -1,43 +1,5 @@
-/*
- *  __   ___   ____  ____  ___   ___
- * ( (` | |_) | |_  | |_  | | \ | | \
- * _)_) |_|   |_|__ |_|__ |_|_/ |_|_/
- *
- * SPEEDD project (www.speedd-project.eu)
- * Machine Learning module
- *
- * Copyright (c) Complex Event Recognition Group (cer.iit.demokritos.gr)
- *
- * NCSR Demokritos
- * Institute of Informatics and Telecommunications
- * Software and Knowledge Engineering Laboratory
- *
- * This program is free software: you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
- * License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with program. If not, see <http://www.gnu.org/licenses/>.
- */
-
 package org.speedd.ml.learners
 
-import org.apache.spark.SparkContext
-import org.apache.spark.sql.SQLContext
+import org.apache.spark.sql.DataFrame
 
-class DataSet {
-
-}
-
-object DataSet {
-
-  def loadFrom(startTime: Long, endTime: Long)(implicit sc: SparkContext, sqlContext: SQLContext): DataSet ={
-    ???
-  }
-}
+case class DataSet(evidence: DataFrame, annotation: DataFrame)
